Routen sortierung, WebApi change
Routen werden jetzt nach BoltRow in der Routenliste sortiert.
Die WebApi gibt nurnoch Routen an die App weiter die in der Halle sind.
In der Routenliste wird nun der Titel je nach Filter geändert.
Rotpunkt wird nun auf der Ampel als Grün angezeigt
Einige kleine Namens und Layout änderungen.

+++ AndroidApp/src/dav/routenbewerter/PieChart.java
@@ -56,7 +56,7 @@ public class PieChart extends Drawable {
 
 			// Start und End Angel bestimmen
 			float endAngle = value_sum == 0 ? 0 : 360 * item.itemValue
-					/ (float) value_sum;
+					/ value_sum;
 			float newStartAngle = startAngle + endAngle;
 
 			//Farbe für den Arc des Items festlegen
